requirements.txt
streamlit
gspread
pandas
pyyaml